Refactor task routes

+++ src/routes/task.py
@@ -1,15 +1,17 @@
-from src import app, db
+from src import app, db, login_manager
 from src.models.task import Task
 from flask import request, jsonify
+from flask_login import login_required, current_user
 
 @app.route("/tasks", methods=["POST"])
+@login_required
 def create_task():
     body = request.get_json()
     title = body["title"]
     description = body["description"]
 
     if title and description:
-        new_task = Task(title=title, description=description)
+        new_task = Task(title=title, description=description, user_owner=current_user.id)
         db.session.add(new_task)
         db.session.commit()
 
@@ -19,8 +21,9 @@ def create_task():
 
 
 @app.route("/tasks", methods=["GET"])
+@login_required
 def get_tasks():
-    tasks_object = Task.query.all()
+    tasks_object = Task.query.filter_by(user_owner=current_user.id).all()
 
     if tasks_object:
         tasks = [task.to_dict() for task in tasks_object]
@@ -30,16 +33,21 @@ def get_tasks():
 
 
 @app.route("/tasks/<int:task_id>", methods=["GET"])
+@login_required
 def get_task(task_id):
     task = Task.query.get(task_id)
 
     if task:
-        return jsonify(task.to_dict())
-
+        if task.user_owner == current_user.id:
+            return jsonify(task.to_dict())
+        
+        return login_manager.unauthorized()
+    
     return jsonify({"error": "Task not found"}), 404
 
 
 @app.route("/tasks/<int:task_id>", methods=["PUT"])
+@login_required
 def update_task(task_id):
     body = request.get_json()
     title = body["title"]
@@ -48,36 +56,45 @@ def update_task(task_id):
     task = Task.query.get(task_id)
 
     if task:
-        task.title = title
-        task.description = description
-        db.session.commit()
+        if task.user_owner == current_user.id:
+            task.title = title
+            task.description = description
+            db.session.commit()
 
-        return jsonify({"message": "Task updated successfully", "id": task.id})
+            return jsonify({"message": "Task updated successfully", "id": task.id})
+        return login_manager.unauthorized()
 
     return jsonify({"error": "Task not found"}), 404
 
 @app.route("/tasks/<int:task_id>/completed", methods=["PATCH"])
+@login_required
 def update_completed_task(task_id):
     task = Task.query.get(task_id)
 
     if task:
-        task.completed = not task.completed
-        db.session.commit()
+        if task.user_owner == current_user.id:
+            task.completed = not task.completed
+            db.session.commit()
 
-        if task.completed:
-            return jsonify({ "message": "Task marked as complete", "id": task.id })
+            if task.completed:
+                return jsonify({ "message": "Task marked as complete", "id": task.id })
+            return jsonify({ "message": "Task marked as incomplete", "id": task.id })
         
-        return jsonify({ "message": "Task marked as incomplete", "id": task.id })
-        
+        return login_manager.unauthorized()
+    
     return jsonify({"error": "Task not found"}), 404
 
 @app.route("/tasks/<int:task_id>", methods=["DELETE"])
+@login_required
 def delete_task(task_id):
     task = Task.query.get(task_id)
 
     if task:
-        db.session.delete(task)
-        db.session.commit()
-        return jsonify({"message": "Task deleted successfully", "id": task_id})
+        if task.user_owner == current_user.id:
+            db.session.delete(task)
+            db.session.commit()
+
+            return jsonify({"message": "Task deleted successfully", "id": task_id})
+        return login_manager.unauthorized()
 
     return jsonify({"error": "Task not found"}), 404
